67- Upgraded Blog/requirements.txt
Bootstrap-Flask==2.3.0
Flask==2.3.2
Flask_CKEditor==0.4.6
Flask-SQLAlchemy==3.1.1
Flask_WTF==1.2.1
WTForms==3.0.1
python-dotenv==1.0.0
